Use powerline for the tmux status line
TIL that python's json parser is utterly retarded; it throws exceptions
at perfectly valid commas at the end of lists. It also does not accept
either /* */ or # for comments.

TIL that tmux apparently has a hardcoded limit on the hardstatus length,
as having too many segments made the status simply not be rendered
despite being valid.